refactor(changelog): share .md serializer, precompute neighbors, drop dead css
Extract renderEntrySectionsMd so the aggregate and per-release markdown
endpoints share one entry serializer (heading level parameterized) instead of
duplicating the format.

Compute same-stream prev/next neighbors once in getReleasePaths and pass them
as props, so detail pages no longer re-read the whole collection per page.

Remove orphaned .cl-anchor/.cl-anchor-tip selectors left after the copy-link
button was replaced by the version-number link.
