+++ site/src/styles/changelog.css
@@ -287,8 +287,8 @@ a.cl-entry-title:hover { color: var(--cl-accent-text); text-decoration: underlin
 .cl-area[hidden], .cl-area-item[hidden], .cl-empty[hidden], .cl-novel[hidden],
 .cl-badge[hidden], .cl-more[hidden], .cl-lang button[hidden] { display: none; }
 
-/* Respect reduced-motion: drop the cosmetic transitions (rail color, anchor
-   fade, tooltip, drawer slide). Functional state changes still apply instantly. */
+/* Respect reduced-motion: drop the cosmetic transitions (rail color, drawer
+   slide). Functional state changes still apply instantly. */
 @media (prefers-reduced-motion: reduce) {
-  .cl-dot, .cl-line, .cl-anchor, .cl-anchor-tip, .cl-rail { transition: none; }
+  .cl-dot, .cl-line, .cl-rail { transition: none; }
 }
